+++ doc/sphinx-guides/source/api/changelog.rst
@@ -10,6 +10,7 @@ API Changelog
 
 New
 ~~~
+- **/api/admin/clearThumbnailFailureFlag**: See :ref:`thumbnail_reset`.
 - **/api/admin/downloadTmpFile**: See :ref:`download-file-from-tmp`.
 
 Changes